Initial version
Let's start with the Tandy "RS 01.00.00" release of the C compiler. The
Dragon release is identical except for the version string in cc1. Oh, not
quite -- the "intercept.a" file from the Dragon disk is intact, while the
one from Tandy's disk 2 in the archives has some odd corruption. I have
used the Dragon release for interrupt.a.

The "ccguide.pdf" documentation available in various places claims this to
be Microware's 1.1 version of the compiler. By looking at the "Differences
between Versions 1.1 and 1.0" section, I believe that this is indeed
version 1.1 because I can verify the FILE struct contains a _bufsiz member
and all of the other changes I know how to verify have been made.

I have included the single-pass "Level 2" version (c.comp) of the compiler
executable in addition to the two-pass (c.pass1, c.pass2) version
distributed by Tandy and Dragon Data. I do not know the origins of the
single-pass version, but I suspect it came from a Gimix or SWTPC system of
some sort.