Add structured help system
It's so much better than the old Help that it's pointless to even consider
the old one. It looks for help files (*.hlp) in /DD/Help, but in the future
I plan to move them to /DD/Sys/Help to get 'em out of the root directory.

That is, once I get it recompiled on CoCo and moved back here. :)