David Betz's XLisp 1.2
XLisp is an object-oriented Lisp written in C. It's apparently pretty rare,
I only found one copy of the OS9 build anywhere on the net.

The source says it's public domain. There is a newer distribution by its
author, apparently maintained, for modern OSes. But this one runs on a
CoCo. :)