Move to Kreider '91 libc as a base
This is, as far as I know, the last version of Carl Kreider's modified
libc (yes, modified, it's not a fully new C library). I did not include all
of the header files included in the "C_Defs" archive on the RTSI site,
because according to the readme included in that archive it wasn't from
Carl anyway and was just all the headers that happened to be on somebody's
hard drive. :)

As such, some of the content in these headers may actually be wrong, and
while that really sucks, let's look at it as an opportunity, shall we?