New Compiler tool source: cpp
Maybe this fell off a truck, but it's an amazing simulation of a Unix port
of the C preprocessor for OS9. And according to the history file, it's way
newer than anything we saw back in the day.

I hope to port this back to (Nitr)OS9/6809 ASAP, maybe it'll be better than
the crappy options we have.