MWSys: Trivial assembly fixes
By comparing disassemblies to what I now know is the original source code
I noticed that some of the "push" and "pull" instructions disassembled
differently from what the source said. And after consultation with
lostwizard, it seems that the disassembly is actually correct (technically
it doesn't matter, because the order is hardcoded into the CPU, but surely
it's better to write the order in which the CPU will actually read/write
the registers.

So let's canonicalize the push/pull.