Jim McDowell's C_PREP version 1.5
This is the very earliest version I can find -- in fact, until yesterday I
thought that version 1.8 was the earliest version still in existence.

It's a mostly drop-in replacement for the Microware c.prep, but with an
ANSI mode triggered by calling it with the argument -c which causes it to
stop emulating c.prep and do something else.