c_prep: 1.8 -> "1.9"
This is Gene Heskett's modification of c_prep 1.8.

Why add the quotes to the version number? Because as far as I know, this
and all future updates are unofficial and amount to minor patches.

The IPatch files are gone now, if you want to replace c.prep you're on your
own. Other than that, most of the changes are the fixing of typos, adding
comments, and adding more debug printfs.

Gene fixed a couple of bugs, too:

in expand() (cp4.c, line 26), Gene changed an "if (!*lpcntr)" string
length check into a null pointer check by removing the *. This is probably
not the complete fix, though it apparently did work. The correct fix is
probably "if (!lpcntr || !*lpcntr), which will catch both an empty string
and a null pointer. It should never have been called with a null pointer
in the first place, though, so that's a second bug present that _wasn't_
fixed. :)

In doundef() (cp1.c, line 79), Gene fixed a bug in the #undef directive
that apparently caused it to be entirely nonfunctional. Gene's fix, which
seems plausible, works around a "feature" of c_prep which puts spaces in
front of and behind every token string. Gene moved a pointer in doundef()
to skip the first char of the token to compare against, allowing the
comparison to succeed (and #undef to function).

He also increased the preprocessor's built-in limits and prevented it from
crashing when they are exceeded, and he apparently disassembled the solve.r
and findstr.r binary files shipped with the preprocessor. I am not entirely
sure why he disassembled solve.r, because I found the source code to it
elsewhere. :)

Once it's up to date, I'll commit the original source to solve.r.