c_prep: "1.9" -> "1.9A"
Moving on to Willard Goosey's mods! Willard made the ANSI special macros
(__FILE__ and so on) expand as string literals instead of simple
replacements, and he... mostly... worked around a Y2K bug in the Kreider
libc's version of asctime().

It looks like Willard noticed the same whitespace issue Gene saw, and to
make __STDC__ expand to 1 added a space on the end of it.

Oh, and he made a Makefile. :)