c_prep: "1.9A" -> "1.9B"
More Willard, I think. In this exciting installment, Mr. Goosey fixes a
couple of bugs! Specifically, c_prep had previously barfed on code using a
Microware extension for what appear to be decimal character literals,
which use "\d[string of numbers]" to represent an ordinal character value.

As well, Willard seems to have fixed another part of the struct member
access tokenization bug, this time when the struct is an element of an
array. That is, foo[1].bar is a single token. If all this seems strange,
that's only because it is -- most of these bugs are caused by c_prep's
crazy reformatting of the source it's processing.

This commit is not precisely as it was originally found in the distribution
lha archive -- I have fixed a couple of important typos and elided some
whitespace changes.

The original README and source comments changed "\d"
to "/d", which is wrong and confusing, and for some unknown reason the
author did some significant whitespace-only changes to cp1.c. Those errors
have been corrected.