c_prep: "1.9B" -> "1.9C"
This is Walter Zambotti's bug fix for c_prep's bug in assuming OS-9 files
are newline-terminated (they aren't; they're newline-separated, which
means that the last line doesn't necessarily have to have a newline at the
end of it).

Well, mostly. I have corrected the documentation, which said that TRUE,
FALSE, and ERROR were not defined by any header -- they're defined by the
Kreider version of <stdio.h>, but not the Microware one. We're using the
Kreider headers, so we don't really need the definitions at all, but since
they are harmless I left the new definitions in.