New Compiler tool source: copt
The amazing simulation continues. This code appears to be similar to a Unix
port of the assembly output optimizer. It is, however, hard to tell since
there really are very few comments.

In the next commit, I should be able to add a makefile and hopefully, a new
working binary that was built on a CoCo and works at least as well as the
original. :)