copt: Woo, it builds!
It still doesn't work quite right, but it builds and runs. I took care of
the warnings and the build failures, and now I got a nifty little "t.opt"
executable in the source dir that runs!

The Makefile turns on DEBUG mode; run the executable with -d to get a ton
of debug output for figuring out what it's doing.

Pulled in and de-ANSIfied an "inst.c" from a different version of the code,
seems to work.

Dunno why it's trying to get ctype.h from rma, though. I don't have (real)
RMA source, so I used the normal clib #include <ctype.h>. Seems to work.