copt: Well, now...
The output is mostly, but not completely, correct. On comparison with the
"test" file, our code was outputting bare lines that contain just "lbra"
and so on, while the other optimizer is outputting the same branches, but
to labels like _$11 -- which exist (but, again, not in our code).

Tracking through the code, I found something which just shouldn't work: the
line:
    return sprintf(lblp, "_$%d",++labnum);
but the function returned a string, not an int like sprintf returns.
In short, there is no way this should ever have worked.

Now we're outputting labels named _, (and branches to them) but something
is killing the $nn part.