copt: fix _$nn labels
Found the fix! Turns out, this part was my fault the whole time. :)

When I used the macros and chcodes array from <ctype.h> and the Kreider
library, it turns out that was the wrong thing to do. RMA has its own
special character classes, and they don't match C's. An RMA identifier can
contain uppercase or lowercase letters, numbers, the special chars "_",
"@", and "." (which RMA puts in the same category as uppercase letters for
some reason), and one character all in its own special category, "$"

It also has bit flags for hex digits (which, for some reason, don't apply
to the lowercase 'a' through 'f', only the uppercase versions) and binary
digits 0 and 1.

I copied ctype.h from /Defs and redefined the macros as best I could. I
had to, somewhat, because the code that spits out the labels uses the
isalnum() macro to figure out what it needs. So, I redefined isalnum() to
include every code in the chcodes2 array that has a nonzero value, and now
it seems to work!

That said, it seems that this version of the optimizer might actually be
older than the one we have -- now that it's working, I can see it's
producing some slightly odd (but working) code. For example, where the
"Tandy" c.opt #4 might produce:

_72 ldb [6,s]
 cmpb #60

this one does:

_72 ldb [6,s]
 sex
 cmpd #60

In fact, it appears that every time an ldb is done, it gets sign-extended.
Probably we need to add something to check the width of the immediate
comparison to determine whether or not to sign-extend B.

The "Tandy" code also uses post-increment indirect addressing, while this
one doesn't. Where the Tandy code emitted an ldb ,u+ instruction, this code
puts out two instructions:

leau 1, u
ldb -1, u

which is just weird, moving u up by one and then assigning b to the value
held in the address pointed at by u minus one. The result is the same, but
why do that? Anyway, the important part is that we have proper source code
to fix it now.