copt: Bwahaha!
I modified the optimizer to allow labels to no longer be immune to
optimizations -- which worked, but it made the labels disappear. Well,
damn, right?

So then I went searching for the right place to put a call to movlab(),
which moves a label from one instruction to another. I am not totally
certain I found the right place, or that it doesn't need some more
conditionals on it to keep a label from moving _up_ the instruction stream.

That is, we can't allow this:

pshs x
_12 pshs y
pshs d

to become this:

_12 pshs d,x,y

Instead, it should be:

pshs x
_12 pshs d,y

Now, having said that, t.opt now produces IDENTICAL OUTPUT! to the original
Microware c.opt, at least for my optim.a test file. I assume that there are
probably other optimizations that had been done before (and will be added),
but none of them affect that file.

Still, progress to a maintainable OS9 C compiler has been made!