copt: Allow instruction references to insns other than the last
The optimizer allowed the special value NULL in replacement patterns to
refer to either the mnemonic or the operand of the instruction being
replaced, but the weird comment in the actions.h suggesting new patterns in
a different format suggests that it might be useful to refer to a mnemonic
or operand from the _previous_ mnemonic/operand, and I agree.

So let's implement it! :)

In replacement patterns, the optimizer now recognizes the special values
NULL, 1, and 2. Possibly in the future, 3, 4, etc. could be supported for
extra lookback, but I am not sure of the utility of that.

NULL and 1 mean the same thing -- the value of the mnemonic or operand being
replaced. 2 means the value of the _previous_ mnemonic, which in our case
is likely to mean an earlier matched mnemonic.

That is: when matching
std <any>
ldd <same>
a replacement pattern of 2, NULL or 2, 2 will cause both instructions to be
replaced with just the "std <any>", having the effect of deleting the ldd.

If you want an argument that does not have an operand to not have one
accidentally inserted, use the null string "", not NULL.