copt: "simple" actions
Somewhat based on the weird header, and because rlink has started giving me
crap about being "out of dref nodes" (which, apparently, has to do with
static structures being too large), I have simplified part of the static
data structure. It's not good enough to handle rules that include anything
other than than two target instructions and one replacement instruction --
but most of the rules meet that standard, and for those that don't we still
have the regular actions[] array (which is now quite small).

the format is as follows:
mnem1, op1, mnem2, op2, replm, replop

The first four form two instructions matched against, while the last two
are the replacement instruction. The array is snarfed in the addsimpl()
function in optim.c (called by optinit()), and action structs are
allocated dynamically and added to the hash table that is _actually_ used to
hold the optimization data.

The second mnemonic (or almost any operand) may be null, meaning "matches
anything", but you should make sure you're matching the right thing. And
possibly more importantly, any pattern where the second mnemonic is null
must be matched against _every_ instruction in the source. The code is
pretty fast, but it can only be so fast, so don't make too many of those.

If the second operand is 1, it matches "anything", but only if it's the
same as the first operand, e.g. std NULL/ldd 1 will match any pair of
std/ldd instructions that use the same operand.

I had to do some hacking to allow null instruction patterns to match
against instructions that might _also_ be matched by other patterns. I was
assisted slightly by the structure of the optimizer; the hash table used
to match rules is keyed to mnemonics, and the hash of null is zero, so if
an instruction is not matched I reset the hash bucket used to the first one
(index zero) and start the search over again. It does slow down matching
slightly, but not overmuch, and it does seem to work now -- but don't hold
your breath. :)