copt: Past parity!
I think we've reached a point where we've reached (and passed) optimization
parity with Microware's Edition #4. This code might be slightly slower, but
the code it _produces_ is slightly smaller and faster than Microware's.

The old c.opt has been renamed totally to c.old, and the new one installed
as c.opt.