cc09: experimental compiler
This appears to be a C cross-compiler targeting OS9/6809. For my next trick,
I shall see about porting it (back?) to the CoCo.

It shows signs of having been a two-pass compiler at one time, but those
days are probably over -- the functions being called for two-pass mode are
just not there any more. The differences would be triggered by defining
the SPLIT symbol, but all it'll really do is show you which functions are
missing. There's also a PASS2 symbol that's only referenced twice.

Interestingly, the compiler has the ability to disable floating-point
variables. If you undefine DOFLOATS, then you can get a smaller compiler
that might be able to build bigger sources, as long as you don't need floats
or doubles. Having said that, a compiler built without DOFLOATS will _never_
be able to recompile the compiler with DOFLOATS enabled, so it doesn't help
us actually get it self-hosting.