Get CC09 compiling on x86-64 Linux
Mostly this is just small changes to APIs on modern systems. The canonical
define for Unix systems back in the day was "unix", not "UNIX", and even
though the more standards-correct version today is "__unix__", I'll leave
it alone for now.

There are also a couple of problems that result PDP-11 Unix having the two
bytes starting at location zero of a process's address space also
containing zero bytes. This is fantastically useful in principle, because
it means you can _safely_ dereference most NULL pointers and just get NULL
or zero back, and you might get unexpected results, but the program won't
crash or do something _really_ weird. But as with other "all the world's a
VAX" conventions from early C, it doesn't hold for all other systems.

For compatibility with PDP-11 Unix, the OS-9 C runtime setup (cstart) sets
up a C program's address space that way too, but other C platforms don't,
which means programs written on early Unix (and OS-9) sometimes segfault
when run on other systems.