CC09: Enable tree printing functionality
For debugging purposes, the compiler has the ability to print out a
representation of how it understands the statements it is processing,
how it decides to optimize a statement, and so on. This is activated by
defining PTREE when compiling it, and then using the -d (debug) argument to
the compiler when it runs.

With a little detective work, I managed to decipher how it actually works,
and reconstruct some idea of what was in the "ckeys" file that was missing
from the source. I have included the file I made to satisfy the compiler:
put it in /usr/local/share/cc09 on a Linux system, or in /dd/lib on an OS9
system if you want to understand how the C compiler looks at your code, or
if you want to help me make the compiler better. :)

Now, this can result in a truly huge quantity of text output, so it may be
wise to use it on, say, a Linux system rather than a CoCo.