CC09: add a more modern Makefile
This is the makefile I use to build the compiler on Linux. It has an
"install" target, which installs the compiler binary into ~/bin, creating
it if necessary, and installs the "ckeys" file as "~/.ckeys".

During compilation, the Makefile alters the location the compiler looks for
the ckeys file to the location you built it for. This absolute location is
hardcoded into the build, so even if it's run by another user it will STILL
look for ckeys in your home directory, because I haven't made the compiler
smart enough to have a config file or anything, or try to read its
environment to find it. If you want to install it for all users, you will
have to change the makefile a bit.

As far as I know it doesn't actually require GNU make, but I don't want to
collide with the OS9 makefile, and changing the filename to "GNUmakefile"
was the easiest way I could think of to prevent that.

Still, now other people should be able to experiment with cc09. :)