Libs/KLibc: Kreider C library
Yeah, I'm just gonna say it: This is the mother lode.

As far as I can tell, this is _exactly_ the source code to the April 1991
("Final") release of Carl Kreider's version of the standard C library for
the Microware C compiler. Seriously, inside the tarball that I received
from Mr. Kreider were (in addition to the stuff I am committing here) ROF
object files that came out of the assembler in 1985 to 1991.

The vast majority of this code is assembly, but interestingly enough some
of the files contain C in the comments -- it looks like someone wrote that
code in C, compiled it with comments included, and then hand-optimized the
resulting assembly...which means at some point, we _should_ be able to use
the C preprocessor to conditionally compile the optimized versions.

Enjoy.