CC09: Get ints working for little-endian systems
Mike Furman helped a lot by narrowing down the basic location of the
"big-endian integers" problem with to the compiler's lexical analyzer,
which narrowed down the search space considerably. The decompiled version
did not really help at all, because it seemed to be fixing the problem in
the wrong place, the code generator.

I originally started by trying to convert the big-endian ints into native
ones (whatever the host endianness), and that actually worked reasonably
well, but it felt really fragile. But then I found the place where numbers
were being parsed in the first place and rewrote that.

The original code was being entirely too clever, but they can be forgiven
because the biggest integer data type they had access to was only 32 bits
wide. We don't have that problem any more, so we can do it in a portable
way -- as long as your system has a "long long" data type. The original
code for making the synthetic long long int is still being used for OS9/OSK
systems, which are both big-endian, so it *should* be OK.

Floating point numbers are probably still garbage -- you probably don't
want to try cross-compiling code with floats or doubles yet.