CC09: Make negative offsets work right
The previous commit fixed many of the weird outputs, but there were many
instances of lines like:

 leas 65530,s

Which is technically correct, since the assembler should turn that into
the right thing...but they really should look like:

 leas -6,s

Well, as it turns out, all the math was correct, it was just a formatting
bug. We were passing a short int to printf(), but telling printf to format
an int using "%d". So it got the bit pattern correct, but didn't know it
was supposed to be negative. printf() has a format string for short, "%hd",
and using it gives us what we expect.