rdump: detect version 1 (RMA) object files
RMA produces object files different from those made by c.asm. In particular,
there are two "extra" bytes at the end. I suspect, however, that they
actually have something to do with another difference that the RDUMP
program that shipped with the Development System reveals. That is, when
_that_ rdump displays info on an .r file, instead of printing

Edition:     0

it displays

Edition/ROF: 0/0 (or 0/1, if the file was emitted from RMA)

By manually dumping the file, I discovered that a version 1 file has a 1
instead of a 0 in the "h_spare" field of the header structure describing
it. So, let's support that change.

This change renames the h_spare field of the object file format header
to h_vers, and when that field is 1 display that fact. At the same time,
let's skip over those extra two bytes so that if we have a library (which
is really just a bunch of ROF files concatenated together) we are then
able to dump the whole library instead of breaking after the first version
1 file we encounter like it always did before.