rdump: dump "common blocks"
I am not entirely sure what a common block is _for_, but that's what those
"extra" two bytes at the end of version 1 object files are doing It's not
a random footer at the end, it's a whole section that happens to usually be
empty!

By groveling through the RMA code, I managed to find where it outputs the
data that goes into the common section, and how to get RMA to produce a
file with that section in it. It turns out there's a directive, called
"common" or "com", that allows you to allocate a number of bytes in one
program section (psect) that will be shared between all psects that use the
same common section with the same name.

Anyway, once I got RMA to produce a file with one, I fed it to the OS-9
Level II Development System's version of "rdump", and it gave me enough
information to understand a bit more of what's going on, including how
to decode it myself.

The simple RMA assembly file

 psect joe,0,0,0,0,joe

foo common 100

joe:
    sta foo+4,pcr
    rts

 ends

is enough to create a common block named "foo" 100 bytes in size, with 1
reference to a location 4 bytes (or 1 word) into it, producing a "common
blocks" symbol table:

The beginning of the symbol table sets up a counter:
	$0001 (1 block)

Immediately following the counter is its name, in a C string:
	$66 $6F $6F $00 ("foo\0")

Immediately following the number is the size of the block:
	$0064 (100 bytes)

Immediately following the size, is another counter, this time for the
count of references to locations within the common block. In our case,
we have 1 reference:
	$0001

Immediately following the counter, we have a reference type/location
byte: $A7, which according to the description in rdump.c, means we
have a relative reference ($80) to a location in the data/code section
($20), and with "CODENT/DIRENT/INIENT" options all set. The native
rdump program describes this byte as referring to an object "in
code/word/pcr"

Following that attribute byte, we have a 16-bit offset: $0002, which
in this case I think describes a distance of two _words_ away, not
bytes.

Anyway, I wrote all of that to say "I can decode version 1 ROF files now".

:)