rdump: formatting, bug fixes
Match the dev system's RDUMP a little more closely. When it outputs a
string that can be either plural or singular, the output is correct for the
number it's outputting. Yes, that might seem a little silly to some people,
but if we're going to _replace_ a program (which I hope to do with this
version, at some point), it ought to be at least as good as the program
it's replacing.

So, I added a "plural()" function. When called with a number 1, it returns
an empty string, but when called with any other number, it returns a string
containing "s". The practical upshot is that you can do:

printf ("I see %d light%s.\n", lights, plural(lights));

and if "lights" is 1, then the program will print "I see 1 light."

Obviously the correct answer is 4, but this works for all (int) n. :)