CC: C compiler executive "2.2.0"
This is my best attempt at a reconstruction of Carl Kreider's compiler
executive, version 2.2.

What I have done is take the oldest version of cc I know about (2.2.1), and
reversed all of the changes that I believe Bill Dickhaus used to make it.

Because of that, this compiler executive does not have support for the
peculiar -r[=filename] option used by the 'cc1' released by Tandy/Dragon.
To get cc to output a ROF file and stop there, you instead use the option
"-r -f=filename" or "-rf=filename"

This version of cc uses c.asm and c.link to assemble and link binaries, and
c.prep/c.comp/c.opt to compile. It does not support the 2-pass version of
the compiler distributed separately by Tandy and Dragon Data for their
OS-9 Level 1 distributions.

There is no build system: compile it with "cc1 cc.c"