CC: "Update" to Bill's 2.2.1
OK, now we get to get back to what we technically started with.

Bill Dickhaus added support for (and made default) the Tandy/Dragon version
of the compiler, and by default made the compiler use the Tandy Level II
Development System (that is, rma and rlink instead of c.asm and c.link).

It also added support for the Make that shipped with the Development
System, by implementing cc1's -r=/path option, and somewhat inexplicably
_removing_ support for outputting .r files the old way with -r -f=/path.