CC: Gene Heskett's "2.5.1" edit.
This, the first of Gene's edits, unconditionally changes the C preprocessor
used by cc, to Jim McDowell's C_PREP, and adds a (hidden -- it's not in the
"usage" output) -E option which enables C_prep's <shudder> trigraph
support.

For his edits, Gene used an editor that inserts tabs for indentation. While
in general I approve of using tabs instead of spaces to indent, it
absolutely needs to be consistent, and this very much was not. Luckily, the
only tabs the editor added were in the lines he edited, so it was easy to
correct.