CC: Gene Heskett's "2.5.2" edit, part 2.
Gene Heskett didn't just add a couple of command line options, he also got
cc compiling with the Ansifront preprocessor. In particular, he had to work
around a "feature" of Ansifront that makes it hash any identifier with a
name longer than the RMA 8 character limit, which causes trouble especially
for the libc function intercept() -- which, you may notice, has a 9-letter
name. :)

In the library, it's known as "intercep", because that's just how long
labels are allowed to be in RMA.

Gene also seems to have fixed an honest-to-goodness bug exposed by
Ansifront that a K&R compiler wouldn't ever notice because of how struct
members work.

In K&R C, struct members are defined as offsets from the beginning of a
struct (that's why they are in a single namespace even today -- no two
struct members can have exactly the same name), so in a K&R C compiler you
can usually get away with having wrongly-named structs as long as they're
the same size, and you can even use the member names for the wrong struct;
the compiler doesn't really know the difference. So, good catch Ansifront!

I'm putting this in separately because they're distinct from the other
changes Gene made.