Add Carl Kreider's debugging library
Apparently most people don't use the C compiler's profiling support. It's
useful when trying to understand which functions are called most often
during a run, which in turn helps figure out what to optimize. A program
compiled with profiling enabled will, on exit, output a list of each
function name that was called, in descending order of how many times that
function was called.

The debugging library, dbg.l, replaces the C library's profiling functions
to, instead of printing a count of all the called functions when the
program exits, prints (to stderr) the name of each function as it's called
and as it returns.

To use it, pass the -P flag to cc and link against /dd/lib/dbg.l _before_
clib(t).l on the command line.