Sed: a port of Unix sed to (Nitr)OS-9
sed is the Unix stream editor, a non-interactive text editor that uses
commands and regular expressions to transform a stream of text from one
representation to another.

Short version: you can do global search and replace on files quickly, as in
the following:

OS9: sed -e s/hello/goodbye/g < hello.txt > goodbye.txt

which will turn every instance of "hello" into "goodbye", using the Shell to
send the file hello.txt as standard input, and redirecting its output to
goodbye.txt.

sed can do a lot more than that, but that's the most common use for it, a
use for which it is admirably qualified.

For more information on sed, Google is your friend or decode the manual
page (sed.man) which I have checked in, but not yet added any tools to
format into helpful text. That's coming, I assure you!