CPP: handle standard #line directives
Our preprocessor is able to deal with a _version_ of the #line directive,
but not one that anything actually uses (to my knowledge).

A correct #line directive goes
#line number-sequence "filename"

but porting Yacc told me we weren't accepting that, and it turns out cpp
was really badly incompatible with real #line directives.

This commit adds support for various #line syntaxes. We accept:

#line num		/* change the compiler's current line number */
#line num filename		/* like we used to accept, but don't drop extensions */
#line num "filename" /* correct */

Potential bugs:
	We don't check that filename exists.
	If filename is a path, the first / will end processing! Don't.

For GitHub: (Closes #8)