CC09: Inconsistent curly braces sometimes bite
My last CC09 commit caused a segfault on Linux, and probably a confused
compiler on OS-9 when _not_ dealing with longs. *blush*

It was because the old coding style doesn't use curly braces consistently,
and I didn't notice that they were missing. Grr.

At some point I want to run the code through 'uncrustify' to clean it up,
but I want the code to stabilize before doing any large-scale reformatting.
I think I also want the toolchain to not break with tabs first too, because
indenting with tabs is good.