CGfx: Makefile updates
Since we don't yet have the ability to link with ROF version 1 libs, we
need to be very careful about how we create them.

Any object file that is intended to go into a library, needs to be version
0, because rlink can't deal with version 1 _for libraries only_. This could
be intentional, but more likely it's just a bug in rlink. Regardless of the
reason, we need to deal with it. Luckily, we have a (native-only) assembler
(c.asm) that ONLY makes version zero ROFs.

That means we (currently) can't make libraries using the cross-compiler,
but that will hopefully change.