Sed: make file arguments work
Turns out, the open() system call doesn't use the same values on Unix and
OS-9. On OS-9, using the value 0 (which, on Unix, means "read only") gives
you an open file descriptor that can't be used for reading OR writing;
maybe it's useful for something else, I don't know.

I did a temporary, dirty hack for portability -- #defining O_RDONLY to the
value of READ from <lowio.h>. Once I finish a new <fcntl.h> header, there
will be a place for real values to go. But this way the code continues to
be buildable on OS-9 and Linux.

I think it's about time to put a binary of sed into the repo. Consider it
"released". :)