CC09: use mkstemp() on more modern systems
After some discussions with David Breeding, I came up with a quick patch
that gets rid of the use of the mktemp() function (which GCC rightfully
complains about, because it's not safe to use on a multi-user system).
Instead, we use mkstemp() and wrap the returned open file descriptor by
calling fdopen().

So, now our temp files are opened safely.

Also, I'm getting rid of the nonstandard five-X string passed to mktemp in
the first place; it's never been standard, and the Kreider lib we use
tolerates, if not requires, the standard six Xes.