Yacc: add modern -o option
Modern yaccs (like bison) have an -o option that causes the outputted parser
to be written to a filename other than "y.tab.c". In most cases, this
causes the rest of the files output to also be defined relative to that
name.

This commit is my version of that option. -o foo (or -ofoo) outputs the
parser source to the file "foo". At the same time, the header (if a header
is requested) will be written to "foo.h", and the info file (if requested)
will be written to "foo.output"

To do the job, I had to do a passable imitation of how getopt() parses
command-line options. I did write a version that actually uses getopt(), and
that works (and is very clean), but it's about 400 bytes larger in size and
that's a lot to spend on command-line parsing on a 6809. I left it in, but
hidden from the compiler with an #if 0 stanza.