Lex: Lexical analyzer generator
Lex is a program which, given a set of patterns and their attached token
names, is able to generate C source implementing a program to recognize
those patterns and take defined actions whenever those patterns are seen.

That is, if you give it a description of the basic structure and keywords of
a language,the program lex generates will be able to turn a stream of text
into 'tokens' that a parser can use as part of an interpreter or compiler.

TLDR; with this and yacc, you can make the start of a compiler.

This is the pristine, untouched Unix source from the 32V release. First, I
will port it to modern Linux, then to (Nitr)OS-9. Again, I renamed Makefile
to GNUmakefile, added the Caldera COPYING file, and the manual page source.