Lex: Linux port (stdio)
Traditionally, stdin/out/err have been implemented in a way that allowed
them to be used to initialize any FILE * variable. When the ANSI committee
standardized C, they decided that this state of affairs was pretty good, but
didn't require it -- they declared that those symbols must be able to be
assigned to any FILE pointer, but not necessarily _statically_.

Relatively recently, GNU libc made a change that made the std* file handles
no longer able to statically initialize variables, so you can't do

FILE *foo = {stdin};

any more on Linux.

Added a STATIC_STDOUT symbol to work the old way -- presumably it's slightly
more efficient -- and didn't define it on Linux. I also unconditionally
made the change to "ncform", which is inserted into outputted lexers,
because there's no way to control that. The more modern Flex has a directive
called "%option stdinit" which enables the old behavior, but we're not Flex.