Lex: Linux build fixes
Apparently, at one time the calloc() function (which is a lot like malloc(),
except it's defined to return memory initialized to zero) actually had a
partner function, cfree(), which was for returning memory allocated using
calloc().

To the best of my knowledge, _nothing_ uses that today, so I will probably
remove the use of cfree() at some point, but right now I think it's best to
change as little of the code as possible. So here's a cfree() macro! :)

Also, in GNUmakefile, let's generate debugging symbols (and 32-bit binaries).