Yacc: Output default declarations for .h files
When we generate a header file, we should include the definition of yylval
and a default declaration for YYSTYPE, so that we don't have to assume
what it will be in the lexer. Most lexers seem to assume that lex will
output this information, so that must come from either later versions or
from Flex. Either way, it makes sense to output that extra info.