Compilers: new headers
<assert.h>:	Provides an ANSI-compatible assert() macro. For those
unaccustomed to assertions, assert(<expr>) declares an expectation that the
expression "expr" must be true at this point in the program, or the program
exits. It's useful for declaring a function's assumptions in a way that the
program will break loudly (instead of silently) when those assumptions are
not satisfied, while also giving you the option to disable the checks
entirely.

If you include <assert.h>, assertions will be enabled. Disable them by
defining the symbol NDEBUG ("Not Debug") either in a "configuration header"
or on the command line.

<ansi.h>: Based on ideas from Minix and the OS-9/68000 headers, this gives
you some macros for writing ANSIsh code. It defines the following macros:
    * _VOID (expands to either "void" or nothing)
    * _VOIDPTR (expands to either "void *" or "char *")
    * _OP() (hides prototypes from pre-ANSI compilers)
    * const, restrict, volatile (only in pre-ANSI mode)

A declaration using <ansi.h> looks like:

_VOID myfunc _OP((const _VOIDPTR aThing, const char *aString, ...));

In a compiler/preprocessor that defines __STDC__, this expands as

void myfunc (const void * aThing, const char *aString, ...);

In a K&R compiler like the McCosh is right now, it expands as

 myfunc ();

which is a perfectly acceptable function declaration in pre-ANSI (actually,
almost any version of) C.
However, do keep in mind that for the two versions of this declaration to
be compatible, C type promotion rules must be followed! A compiler that
doesn't know the passed-in types promotes char and float arguments to int
and double, respectively; so no char or float function arguments (pointers
to those are OK).

The same is true of variadic function arguments in modern C too, by the way;
you can't pass a bare float to printf() without it becoming a double.

<stddef.h>: ANSI-compatible stddef.h, declaring NULL, size_t, ssize_t,
ptrdiff_t, and the offsetof() macro, which expands to the offset of a
struct member from the start of the struct itself. It's often useful for
constructing pointers to struct members, or figuring out how much data to
read into a variable-sized struct variable.

Say you want to read only the "inode" information for a file and not its
whole segment list; you can malloc() only enough memory to hold the first
part of the struct, with

myvar = malloc (offsetof(struct fildes, fdseg));

There are other uses (many of them obscure), but I won't get into them here.