CPP: fix __FILE__
The ANSI/ISO built-in __FILE__ macro was not working correctly; instead of
reporting the current file, it seems to always report the _previous_ file.
I thought it had been something I did when I added support for the ANSI/ISO
#line directive, so I added a bunch of debug code. But no, that wasn't it.

Turns out, it was just a simple mistake by an earlier unknown coder.

I imagine that before __FILE__ was implemented in the preprocessor, whenever
the current file changed the code would do something like:
	strcpy(filename, theNewName);
so that the "#line" output sent to the compiler would have the right names.
When they started making the preprocessor more ANSI-compatible and added
__FILE__, they needed to cause it to be output as a string literal, which
means sprintf and not strcpy.

In this case, that meant adding a new function, setfile(), which takes
a string containing the new filename. It does a
sprintf(filestring, "\"%s\"", thefilename);
and then assigns filename to be thefilename and returns it. Perfectly done,
everything's correct and will work fine.

The problem came when it was hooked up. When the filename changes, now the
program calls:
	strcpy(setfile(filename), theNewName);

At first glance, this seems strange, but understanding that setfile()
returns the pointer it's given makes you think it's going to work, because
strcpy() is going to be writing to the pointer returned by setfile().
Unfortunately, when this code runs, what's actually happening is that
setfile() is being called with the old name, setting the __FILE__ string
to that old name, and then strcpy updates the name to the new one... so it
winds up being that __FILE__ is always wrong if you #include a header.

Oops. :)

The correct process is to do it this way:
	setfile(strcpy(filename, theNewName));
Looks almost the same, but this way works.